Some renaming and cleanup in the Android audio code
